fix(fiche-patient): corrections revue US-2632 (découplage types + doc)
Suite revue (code-reviewer + healthcare-security-auditor) — GO, corrections :
- M1 : types de vue (GlycemiaView/TreatmentView/SlotCoverage/DocumentItem…)
  relocalisés dans un module neutre `components/diabeo/patient/patient-record-
  views.ts` ; les builders serveur (glycemia/treatment/document-view) les
  ré-exportent → sens de dépendance app→components, PatientRecord autoportant
  (plus d'import du dossier de route).
- M2 : commentaire d'en-tête nuancé (PatientContextBar construit encore des liens
  porteurs d'id patient → contrat opaque traité en US-2633).
- L1 : `documentHref ?? ""` documenté (branche défensive inatteignable).
- L3 : test renommé patient-detail-client → patient-record.test.tsx.
- Doc US-2632 enrichie (implémentation + reste à venir).

tsc/eslint/build OK, 3439 tests verts.
